+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+## 1.1.0
+* ADD: Support of 10000 records
+* ADD: Max number of bins was increased from 100 to 5000
+* ADD: New option for setting of start and end points for X axis
+
 ## 1.0.2
 * Fixed console errors when chart contains only 1 bar